# frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "utoipa-axum",
		"description": "Utoipa's axum bindings for seamless integration for the two",
		"contact": {
			"name": "Juha Kukkonen",
			"email": "juha7kukkonen@gmail.com"
		},
		"license": {
			"name": "MIT OR Apache-2.0",
			"identifier": "MIT OR Apache-2.0"
		},
		"version": "0.2.0"
	},
	"paths": {
		"/account": {
			"get": {
				"tags": ["Account"],
				"summary": "Return user object",
				"operationId": "get_user",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserResponseDTO"
								}
							}
						}
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
...
		"/auth/logout-all": {
			"post": {
				"tags": ["Auth"],
				"summary": "Invalidate all refresh tokens",
				"operationId": "logout_all",
				"responses": {
					"200": {
						"description": "All sessions logged out successfully"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
		"/auth/logout-session/{jti}": {
			"post": {
				"tags": ["Auth"],
				"summary": "Revoke refresh token with jti passed to path parameters",
				"operationId": "logout_session",
				"parameters": [
					{
						"name": "jti",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Session logged out successfully"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
		"/auth/refresh": {
			"post": {
				"tags": ["Auth"],
				"summary": "Return new access and refresh tokens",
				"operationId": "refresh",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RefreshResponse"
								}
							}
						}
					},
					"401": {
						"description": "Invalid or expired refresh token"
					},
					"403": {
						"description": "`email_missing`: the account has no email and may not sign in; all its sessions are revoked"
					}
				},
				"security": [
					{
						"CookieAuth": []
					}
				]
			}
		},
		"/auth/sessions": {
			"get": {
				"tags": ["Auth"],
				"summary": "Return information for all active sessions (refresh tokens)",
				"operationId": "get_sessions",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/SessionInfo"
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [
					{
						"BearerAuth": []
					},
					{
						"CookieAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"GetUserResponseDTO": {
				"type": "object",
				"required": [
					"id",
					"username",
					"email",
					"role",
					"email_verified",
					"profile_complete"
				],
				"properties": {
					"email": {
						"type": "string"
					},
					"email_verified": {
						"type": "boolean"
					},
					"first_name": {
						"type": ["string", "null"]
					},
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"last_name": {
						"type": ["string", "null"]
					},
					"patronymic": {
						"type": ["string", "null"]
					},
					"profile_complete": {
						"type": "boolean"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"username": {
						"type": "string"
					}
				}
			},
...
			"RefreshResponse": {
				"type": "object",
				"required": ["access_token"],
				"properties": {
					"access_token": {
						"type": "string"
					}
				}
			},
...
			"SessionInfo": {
				"type": "object",
				"required": [
					"jti",
					"is_current",
					"device_id",
					"device_label",
					"last_used",
					"issued_at"
				],
				"properties": {
					"device_id": {
						"type": "string"
					},
					"device_label": {
						"type": "string",
						"description": "Human-friendly device label, e.g. \"Chrome on Windows\"."
					},
					"ip": {
						"type": ["string", "null"]
					},
					"is_current": {
						"type": "boolean"
					},
					"issued_at": {
						"type": "string",
						"format": "date-time"
					},
					"jti": {
						"type": "string",
						"format": "uuid"
					},
					"last_used": {
						"type": "string",
						"format": "date-time"
					},
					"user_agent": {
						"type": ["string", "null"]
					}
				}
			},
...
			"UserRole": {
				"type": "string",
				"enum": ["Student", "Teacher", "Admin"]
			},
...
		}
	}
}